Block: confining ash prototype
Prevents undead from walking over it. At the moment it uses normal ash
to place the block, but it will be a custom item in the future, based on
ash, and you'll be able to select which entities it will block
(vampires, werewolves, undead, demons, or all of the above). It will
replace the ritual of confinement since ritual stacking will not be a
thing (cannot use more than 1 ritual at the same time). In normal and
hard mode it will act like doors, zombies will try and destroy it